requirements.txt
feedparser
anthropic
openai
kokoro-onnx
pydub
soundfile
PyYAML
Jinja2
python-dotenv
